resizer/requirements.txt
pika==1.1.0
Pillow-SIMD==7.0.0.post3
# Pillow==7.1.1 in case of Pillow-SIMD installation errors on Windows

